# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce Core API Documentation",
        "description": "API documentation for Laravel E-Commerce Core project including Auth, Products, Categories, Cart, Orders, Coupons, Reviews/Comments, Addresses, and Admin Management.",
        "contact": {
            "name": "API Support Team",
            "email": "admin@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Danh sách đơn hàng của tôi",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Danh sách đơn hàng phân trang"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Tạo đơn hàng mới (Checkout)",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "customer_email",
                                    "customer_phone",
                                    "shipping_address",
                                    "items"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "customer@example.com"
                                    },
                                    "customer_phone": {
                                        "type": "string",
                                        "example": "0987654321"
                                    },
                                    "shipping_address": {
                                        "type": "string",
                                        "example": "123 Nguyễn Huệ, Q.1, TP.HCM"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2,
                                                    "minimum": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "product_id": 1,
                                                "quantity": 2
                                            }
                                        ],
                                        "minItems": 1
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "cod",
                                        "nullable": true,
                                        "enum": [
                                            "cod",
                                            "bank_transfer"
                                        ]
                                    },
                                    "coupon_code": {
                                        "type": "string",
                                        "example": "SUMMER2026",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Giao trong giờ hành chính",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo đơn hàng thành công"
                    },
                    "422": {
                        "description": "Dữ liệu đơn hàng không hợp lệ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Xem chi tiết đơn hàng",
                "operationId": "8114ac185cc1bf301d8862617e497098",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết đơn hàng"
                    },
                    "404": {
                        "description": "Đơn hàng không tồn tại hoặc không thuộc về người dùng"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Hủy đơn hàng",
                "operationId": "6b3a53e3f3f98b4e5da616fa85e4f37d",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID đơn hàng",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hủy thành công"
                    },
                    "400": {
                        "description": "Đơn hàng không thể hủy ở trạng thái hiện tại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {your_token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}